+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ### Added:
 - 
 ### Fixed:
-- Fixed levelz compat
-- Fixed tooltipfix compat
+- Fixed offhand usage
 ### Changed:
 -